+++ src/main/java/ing/yisus/apihotelera/Controllers/BillController.java
@@ -1,5 +1,6 @@
 package ing.yisus.apihotelera.Controllers;
 
+import ing.yisus.apihotelera.Exeption.ResourceNotFoundExeption;
 import ing.yisus.apihotelera.Persistence.BillEntity;
 import ing.yisus.apihotelera.service.BillService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,16 +22,25 @@ public class BillController {
 
     @GetMapping("get/{id}")
     public BillEntity obtenerFacturaPorId(@PathVariable Integer id){
+        if(billService.getBillById(id) == null){
+            throw new ResourceNotFoundExeption("get","id", id);
+        }
         return billService.getBillById(id);
     }
 
     @DeleteMapping("/delete/{id}")
     public void eliminarFactura(@PathVariable Integer id){
+        if(billService.getBillById(id) == null){
+            throw new ResourceNotFoundExeption("get","id", id);
+        }
         billService.deleteBill(id);
     }
 
     @PutMapping("/update/{id}")
     public void actualizarFactura(@RequestBody BillEntity bill, @PathVariable Integer id){
+        if(billService.getBillById(id) == null){
+            throw new ResourceNotFoundExeption("get","id", id);
+        }
         billService.updateBill(bill);
     }
